+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Next
 
+### Added
+
+* Linux support
+* `ScrappableURL.userInfo` property for storing values or objects related to this scrappable URL
+
 ## [0.1.0](https://github.com/DenTelezhkin/Swarm/releases/tag/0.1.0)
 
 * Initial release